+++ docs/HISTORY.rst
@@ -16,6 +16,7 @@ Change Log
 - now_is_between() now supports specification of a timethat can be used for testing rather than using the current time
 - Added the ability to use lambdas for filtering state callbacks - contributed by `David Greenberg <https://github.com/dgrnbrg>`__
 - Added the ability to support multiple input constraints - contributed by `Matt O <https://github.com/SirIndubitable>`__
+- Added the --configfile parameter
 
 **Fixes**